NeRF_LiDAR/zipnerf/requirements.txt
numpy
torch==2.0.0
absl_py
accelerate
gin_config
imageio
matplotlib
mediapy
opencv_contrib_python
opencv_python
Pillow
plyfile
rawpy
scipy
scikit-image
tensorboardX
tqdm